requirements.txt
pdfplumber
python-docx
pandas
openpyxl
sentence-transformers
faiss-cpu
openai
langchain
fastapi
uvicorn
